requirements.txt
adal==1.2.4
aniso8601==8.0.0
applicationinsights==0.11.9
argcomplete==1.12.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astroid==2.4.2
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
azure-cli-core==2.11.1
azure-cli-telemetry==1.0.6
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
azure-core==1.8.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-core==1.2.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.4.0
azureml-automl-core==1.12.0
azureml-core==1.12.0.post1
azureml-dataprep==2.0.7
azureml-dataprep-native==20.0.2
azureml-dataset-runtime==1.12.0
azureml-pipeline==1.12.0
azureml-pipeline-core==1.12.0
azureml-pipeline-steps==1.12.0
azureml-sdk==1.12.0
azureml-telemetry==1.12.0
azureml-train==1.12.0
azureml-train-automl-client==1.12.0.post1
azureml-train-core==1.12.0
azureml-train-restclients-hyperdrive==1.12.0
Babel==2.8.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach==3.1.5
blinker==1.4
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1598352799320/work
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
dnspython==2.0.0
docker==4.3.1
dotnetcore2==2.1.14
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-Principal==0.4.0
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fusepy==3.0.1
humanfriendly==8.2
idna==2.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598353543543/work
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.4.2
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1596472767194/work
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
knack==0.7.2
lazy-object-proxy==1.4.3
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mongoengine==0.20.0
msal==1.0.0
msal-extensions==0.1.3
msrest==0.6.18
msrestazure==0.6.4
natsort==7.0.1
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
ndg-httpsclient==0.5.1
notebook @ file:///C:/ci/notebook_1596837179121/work
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
passlib==1.7.3
pathspec==0.8.0
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
pkginfo==1.5.0.1
portalocker==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pyarrow==1.0.1
pyasn1==0.4.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==3.1.2
Send2Trash==1.5.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0
